[CLEAN] Cut features in components & hooks
