--- hooks/use-clock.ts
@@ -1,25 +1,0 @@
-import { useEffect, useState } from 'react'
-
-function useClock() {
-    const [clock, setClock] = useState(0);
-
-    useEffect(() => {
-        let timeOutId: NodeJS.Timeout;
-        const clockRunner = () => {
-            timeOutId = setTimeout(() => {
-                setClock(new Date().getTime());
-                clockRunner();
-            }, 1000);
-        }
-        clockRunner();
-
-        // Removes the song on unmount
-        return () => {
-            if (timeOutId) clearTimeout(timeOutId);
-        }
-    }, []);
-
-    return clock;
-}
-
-export default useClock;
