+++ tailwind.config.ts
@@ -26,6 +26,7 @@ const config: Config = {
     },
     extend: {
       colors: {
+        'spotify-green': '#22C55E',
         border: "hsl(var(--border))",
         input: "hsl(var(--input))",
         ring: "hsl(var(--ring))",
